+++ src/components/month/calendarMonthView.component.ts
@@ -8,7 +8,8 @@ import {
   OnInit,
   OnDestroy,
   LOCALE_ID,
-  Inject
+  Inject,
+  TemplateRef
 } from '@angular/core';
 import {
   CalendarEvent,
@@ -45,17 +46,11 @@ import { CalendarEventTimesChangedEvent } from '../../interfaces/calendarEventTi
   selector: 'mwl-calendar-month-view',
   template: `
     <div class="cal-month-view">
-      <div class="cal-cell-row cal-header">
-        <div
-          class="cal-cell"
-          *ngFor="let header of columnHeaders"
-          [class.cal-past]="header.isPast"
-          [class.cal-today]="header.isToday"
-          [class.cal-future]="header.isFuture"
-          [class.cal-weekend]="header.isWeekend">
-          {{ header.date | calendarDate:'monthViewColumnHeader':locale }}
-        </div>
-      </div>
+      <mwl-calendar-month-view-header
+        [days]="columnHeaders"
+        [locale]="locale"
+        [customTemplate]="headerTemplate">
+      </mwl-calendar-month-view-header>
       <div class="cal-days">
         <div *ngFor="let rowIndex of view.rowOffsets">
           <div class="cal-cell-row">
@@ -66,6 +61,7 @@ import { CalendarEventTimesChangedEvent } from '../../interfaces/calendarEventTi
               [openDay]="openDay"
               [locale]="locale"
               [tooltipPlacement]="tooltipPlacement"
+              [customTemplate]="cellTemplate"
               (click)="dayClicked.emit({day: day})"
               (highlightDay)="toggleDayHighlight($event.event, true)"
               (unhighlightDay)="toggleDayHighlight($event.event, false)"
@@ -79,6 +75,7 @@ import { CalendarEventTimesChangedEvent } from '../../interfaces/calendarEventTi
           <mwl-calendar-open-day-events
             [isOpen]="openRowIndex === rowIndex"
             [events]="openDay?.events"
+            [customTemplate]="openDayEventsTemplate"
             (eventClicked)="eventClicked.emit({event: $event.event})">
           </mwl-calendar-open-day-events>
         </div>
@@ -133,6 +130,21 @@ export class CalendarMonthViewComponent implements OnChanges, OnInit, OnDestroy 
    * The start number of the week
    */
   @Input() weekStartsOn: number;
+
+  /**
+   * A custom template to use to replace the header
+   */
+  @Input() headerTemplate: TemplateRef<any>;
+
+  /**
+   * A custom template to use to replace the day cell
+   */
+  @Input() cellTemplate: TemplateRef<any>;
+
+  /**
+   * A custom template to use to replace the day cell
+   */
+  @Input() openDayEventsTemplate: TemplateRef<any>;
 
   /**
    * Called when the day cell is clicked